Copy oxygen-icons statically
We don't use any icons which are not 16x16 or 64x64, so including a 368MB submodule is simply irrational, complicates installation and development process. Since the icons are already sugoi, it is also very unlikely that we will need to bump their version in the nearest frature. This being said, I'm removing this submodule and replacing it with actual files copied from repo.